+++ packages/org/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @docube/org
 
+## 0.4.0
+
+### Minor Changes
+
+- Auto-generate slugs in meta fields from original filenames
+
+### Patch Changes
+
+- Updated dependencies
+  - @docube/common@0.4.0
+
 ## 0.3.0
 
 ### Minor Changes